Reorganized Constraints for Readability
